requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==4.0.1
django-autoslug==1.9.8
django-bootstrap4==21.2
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-js-asset==1.2.2
gunicorn==20.1.0
Pillow==9.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
whitenoise==5.3.0
